tests/submission/W3C/domhtml/anchor03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Anchor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-a-coords" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<p>
<a coords="0,0,100,100" href="../images/yellow.png">View Submit Button</a>
</p>


<script>
var domtest = async_test("The coords IDL attribute of the a element must reflect the respective content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcoords;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("a");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vcoords = testNode.coords;

      assert_equals(vcoords,"0,0,100,100");
       
    domtest.done();
}));

</script>
<p>The coords IDL attribute of the a element must reflect the respective content attribute of the same name.</p>
<div id="log"></div>
</body></html>
